pyproject.toml
[project]
name = "mau-card"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "redis>=6.0.0,<7.0.0",
    "fastapi>=0.115.12,<0.116.0",
    "uvicorn>=0.34.2,<0.35.0",
    "loguru>=0.7.3,<0.8.0",
    "mau",
    "pillow>=11.2.1",
]

[tool.uv.sources]
mau = { git = "https://github.com/miroqru/mauno" }

[dependency-groups]
dev = [
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
